requirements.txt
python-decouple
getpass
requests
windows-curses
